requirements.txt
astroid==2.1.0
autopep8==1.4.3
gitdb2==2.0.5
GitPython==2.1.11
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.7.6.1
pycodestyle==2.4.0
pylint==2.2.2
rope==0.11.0
six==1.12.0
smmap2==2.0.5
SQLAlchemy==1.2.15
SQLAlchemy-Utils==0.33.10
typed-ast==1.1.1
wrapt==1.10.11
